PWM RGB

+++ T-Bird-Project/T-Bird-Project/rgb.h
@@ -14,7 +14,7 @@
 #ifndef RGB_H_
 #define RGB_H_
 
-void rgb_Show(float r, float g, float b, uint8_t brightness);
+void rgb_Show(uint8_t r, uint8_t g, uint8_t b, uint8_t brightness);
 
 void rgb_Rainbow();
